(Pre-UI) Reworking of AppDelegate and (Both) Added UI Tests for the new password view
