+++ Patched Sur - For Catalina/Assets.xcassets/PSIcon.imageset/Contents.json
@@ -1,0 +1,22 @@
+{
+  "images" : [
+    {
+      "filename" : "icon_64x64.png",
+      "idiom" : "universal",
+      "scale" : "1x"
+    },
+    {
+      "filename" : "icon_128x128.png",
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "idiom" : "universal",
+      "scale" : "3x"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
