+++ Patched Sur - For Catalina/CoreCode/main.swift
